Move jsobfu to a mixin
