+++ lib/msf/core/exploit/remote/browser_exploit_server.rb
@@ -4,6 +4,7 @@ require 'erb'
 require 'cgi'
 require 'date'
 require 'rex/exploitation/js'
+require 'msf/core/exploit/jsobfu'
 
 ###
 #
@@ -17,6 +18,7 @@ module Msf
 
     include Msf::Exploit::Remote::HttpServer::HTML
     include Msf::Exploit::RopDb
+    include Msf::Exploit::JSObfu
 
     # this must be static between runs, otherwise the older cookies will be ignored
     DEFAULT_COOKIE_NAME = '__ua'
@@ -82,8 +84,7 @@ module Msf
 
       register_advanced_options([
         OptString.new('CookieName', [false,  "The name of the tracking cookie", DEFAULT_COOKIE_NAME]),
-        OptString.new('CookieExpiration', [false,  "Cookie expiration in years (blank=expire on exit)"]),
-        OptInt.new('JsObfuscate', [false, "Number of times to obfuscate JavaScript", 0])
+        OptString.new('CookieExpiration', [false,  "Cookie expiration in years (blank=expire on exit)"])
       ], Exploit::Remote::BrowserExploitServer)
     end
 
@@ -552,21 +553,6 @@ module Msf
       platform = platform.gsub(/^Microsoft Windows$/, 'Windows')
 
       regenerate_payload(cli, platform, arch).encoded
-    end
-
-    #
-    # Returns an JSObfu object. A wrapper of ::Rex::Exploitation::JSObfu.new(js).obfuscate
-    #
-    # @param js [String] JavaScript code
-    # @param opts [Hash] obfuscation options
-    #    * :iterations [FixNum] Number of times to obfuscate
-    # @return [::Rex::Exploitation::JSObfu]
-    #
-    def js_obfuscate(js, opts={})
-      iterations = (opts[:iterations] || datastore['JsObfuscate']).to_i
-      obfu = ::Rex::Exploitation::JSObfu.new(js)
-      obfu.obfuscate(:iterations=>iterations)
-      obfu
     end
 
   end
