+++ spec/lib/msf/core/exploit/remote/browser_exploit_server_spec.rb
@@ -298,54 +298,4 @@ describe Msf::Exploit::Remote::BrowserExploitServer do
     end
   end
 
-  describe '#js_obfuscate' do
-    let (:js) do
-      %Q|alert("hello, world");|
-    end
-
-    let(:default_jsobfuscate) do
-      0
-    end
-
-    before do
-      subject.datastore['JsObfuscate'] = default_jsobfuscate
-    end
-
-    context 'when iteration is set' do
-      it 'returns a ::Rex::Exploitation::JSObfu object' do
-        opts = {:iterations=>0}
-        obj = server.js_obfuscate(js, opts)
-        expect(obj).to be_kind_of(::Rex::Exploitation::JSObfu)
-      end
-
-      it 'does not obfuscate if iteration is 0' do
-        opts = {:iterations=>0}
-        obj = server.js_obfuscate(js, opts)
-        expect(obj.to_s).to include js
-      end
-
-      it 'obfuscates if iteration is 1' do
-        opts = {:iterations=>1}
-        obj = server.js_obfuscate(js, opts)
-        expect(obj.to_s).not_to include js
-      end
-    end
-
-    context 'when iteration is nil' do
-      let (:opts) do
-        {:iterations=>nil}
-      end
-
-      it 'should return a ::Rex::Exploitation::JSObfu object' do
-        obj = server.js_obfuscate(js, opts)
-        expect(obj).to be_kind_of(::Rex::Exploitation::JSObfu)
-      end
-
-      it 'should not obfuscate' do
-        obj = server.js_obfuscate(js, opts)
-        expect(obj.to_s).to include(js)
-      end
-    end
-  end
-
 end
